Fix bridge compilation: type mismatches, missing Debug, format macro
- Cast MAX_DISPUTE_REASON_LENGTH to u32 for Soroban String::len() comparison
- Add Debug derive to BridgeStatus and BridgeDirection for assert_eq!
- Import Env from soroban_sdk root (private in testutils in v22)
- Add extern crate alloc + use alloc::format for format! macro in no_std

+++ contracts/bridge/src/contract.rs
@@ -836,7 +836,10 @@ impl BridgeGateway {
         disputer.require_auth();
 
         assert!(!reason.is_empty(), "Reason required");
-        assert!(reason.len() <= MAX_DISPUTE_REASON_LENGTH, "Reason too long");
+        assert!(
+            reason.len() <= MAX_DISPUTE_REASON_LENGTH as u32,
+            "Reason too long"
+        );
 
         let tx = storage::get_transaction(&env, tx_id).expect("Transaction not found");
